social_media/requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
gunicorn
Pillow
drf-yasg
drf_spectacular
